exec command norm対応ファイル分割
