+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tterao <tterao@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/25 11:01:17 by hsawamur          #+#    #+#              #
-#    Updated: 2023/08/24 13:27:21 by tterao           ###   ########.fr        #
+#    Updated: 2023/08/24 16:52:56 by tterao           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -85,7 +85,11 @@ EXEC_DIR = exec
 SRCS += $(SRCS_DIR)/$(EXEC_DIR)/exec.c \
 		$(SRCS_DIR)/$(EXEC_DIR)/exec_debug.c \
 	    $(SRCS_DIR)/$(EXEC_DIR)/exec_make_filepath.c	\
-	    $(SRCS_DIR)/$(EXEC_DIR)/exec_get_filepath.c
+	    $(SRCS_DIR)/$(EXEC_DIR)/exec_get_filepath.c	\
+	    $(SRCS_DIR)/$(EXEC_DIR)/exec_child_process.c	\
+	    $(SRCS_DIR)/$(EXEC_DIR)/exec_do_redirection.c	\
+	    $(SRCS_DIR)/$(EXEC_DIR)/exec_fork.c	\
+	    $(SRCS_DIR)/$(EXEC_DIR)/exec_wait_child_process.c
 
 LIBRARY_DIR = library
 LIBRARY_AFILE = $(LIBRARY_DIR)/library.a
