+++ includes/exec_command.h
@@ -6,7 +6,7 @@
 /*   By: tterao <tterao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 15:49:26 by hsawamur          #+#    #+#             */
-/*   Updated: 2023/08/23 19:17:25 by tterao           ###   ########.fr       */
+/*   Updated: 2023/08/24 16:57:43 by tterao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,6 @@ enum e_pipefd
  * @param operator 構文木nodeの一つ上のnodeから引き継ぐoperator
  * @param d 環境変数と終了ステータス
  */
-// void command_execution(t_ast *node, t_operator operator, t_data * d);
 void	exec_command(t_ast *node, t_operator operator, t_data *d);
 
 /**
@@ -61,10 +60,9 @@ void	exec_command(t_ast *node, t_operator operator, t_data *d);
  * @param d 環境変数と終了ステータス
  * @return true すべてのredirectionが問題なく成功した場合、trueを返す
  * @return false redirectionを失敗したタイミングで、この関数の処理を終了し、falseを返す
- * //どういうタイミングで失敗するのか？
  *
  */
-// bool exec_do_redirection(t_ast *node, t_data *d);
+bool exec_do_redirection(t_ast *node, t_data *d);
 
 /**
  * @brief この関数はforkを実行し、子プロセスを生成する。
@@ -74,7 +72,7 @@ void	exec_command(t_ast *node, t_operator operator, t_data *d);
  * @param node 構文木のnode
  * @param d 環境変数と終了ステータス
  */
-// void exec_fork(t_ast *node, t_data *d);
+void exec_fork(t_ast *node, t_data *d);
 
 /**
  * @brief この関数はforkを実行し、子プロセスを生成する。親プロセスは子プロセスの実行結果を受け取る。
@@ -127,7 +125,7 @@ void exec_pipe(t_ast *node, t_data *d);
  * @param pipefd pipeがない場合は、NULLが与えられる
  * @param d 環境変数と終了ステータス
  */
-// void exec_child_process(t_ast *node, int *pipefd, t_data *d);
+void exec_child_process(t_ast *node, int *pipefd, t_data *d);
 
 /**
  * 	@brief この関数は、commandのpathを取得する。
@@ -148,7 +146,7 @@ char *exec_make_filepath(t_ast *node, t_data *d);
  * @param node 構文木のnode
  * @return char** コマンド実行の二次元配列（argv）
  */
-// char **exec_make_command_array(t_ast *node);
+char **exec_make_command_array(t_ast *node);
 
 /**
  * @brief この関数は、第一引数で与えられたnodeより下のnodeの子プロセスを待ち、終了ステータスを取得する。
@@ -158,7 +156,7 @@ char *exec_make_filepath(t_ast *node, t_data *d);
  *
  * @param node 構文木のnode
  */
-// void exec_wait_child_process(t_ast *node, t_data *d);
+void exec_wait_child_process(t_ast *node, t_data *d);
 
 /**
  * @brief この関数は、コマンドがbuiltinか判定する
@@ -170,7 +168,7 @@ char *exec_make_filepath(t_ast *node, t_data *d);
  * @return true builtinの場合、trueを返す
  * @return false builtinde出ない場合、falseを返す
  */
-// bool exec_is_builtin(t_ast *node);
+bool exec_is_builtin(t_ast *node);
 
 // void	debug_printf_double_arr(char **arr);
